Allow headers to be passed in to underlying Axios
