+++ src/test/suite/extension.test.ts
@@ -5,7 +5,7 @@ import * as assert from "assert";
 import * as fs from "fs";
 import * as path from "path";
 import { ExtensionContext, extensions, Uri, window, CancellationTokenSource } from "vscode";
-import IFileDownloader from "../../IFileDownloader";
+import IFileDownloader, { FileDownloadSettings } from "../../IFileDownloader";
 import { ErrorUtils } from "../../utility/Errors";
 import { rimrafAsync } from "../../utility/FileSystem";
 
@@ -18,6 +18,8 @@ const MockExtensionContext = { globalStoragePath: MockGlobalStoragePath, globalS
 
 const TestDownloadUri = Uri.parse(`https://www.w3.org/WAI/ER/tests/xhtml/testfiles/resources/pdf/dummy.pdf`);
 const TestDownloadFilename = `test.pdf`;
+const TestDownloadUriWithSettings = Uri.parse(`https://api.github.com/repos/Azure/apiops/releases/assets/120693194`);
+const TestDownloadFilenameWithSettings = `extractor.linux-x64.exe`;
 
 suite(`Integration Tests`, () => {
     window.showInformationMessage(`Start all tests.`);
@@ -299,5 +301,24 @@ suite(`Integration Tests`, () => {
         for (const item of downloadedItems) {
             assert(!item.fsPath.includes(`50MB.zip`));
         }
+    });
+
+    test(`Simple download from GitHub`, async () => {
+        const settings: FileDownloadSettings = {
+            timeoutInMs: 20000,
+            // eslint-disable-next-line @typescript-eslint/naming-convention
+            headers: {"Accept": `application/octet-stream`, "Content-Type": `application/octet-stream`}
+        };
+
+        assert(
+            await fileDownloader.downloadFile(
+                TestDownloadUriWithSettings,
+                TestDownloadFilenameWithSettings,
+                MockExtensionContext,
+                undefined,
+                undefined,
+                settings
+            )
+        );
     });
 });
