+++ src/test/suite/index.ts
@@ -10,7 +10,7 @@ export const run = (): Promise<void> => {
     const mocha = new Mocha({
         ui: `tdd`,
         color: true,
-        timeout: 10000
+        timeout: 20000
     });
 
     const testsRoot = path.resolve(__dirname, `..`);
